refactor(local-mode): one isPairedLockfileEntry classifier shared by all three guardian-token hosts
